Former-commit-id: be1d4a6fe87b8fd725a3aee9fd45ced8b235b98f

+++ WhirlyGlobeSrc/WhirlyGlobeLib/src/MaplyPinchDelegate.mm
@@ -63,12 +63,12 @@ using namespace WhirlyKit;
 }
 
 // Bounds check on a single point
-- (bool)withinBounds:(Point3f &)loc view:(UIView *)view renderer:(WhirlyKitSceneRendererES *)sceneRender
+- (bool)withinBounds:(Point3d &)loc view:(UIView *)view renderer:(WhirlyKitSceneRendererES *)sceneRender
 {
     if (bounds.empty())
         return true;
     
-    Eigen::Matrix4f fullMatrix = [mapView calcFullMatrix];
+    Eigen::Matrix4d fullMatrix = [mapView calcFullMatrix];
     
     // The corners of the view should be within the bounds
     CGPoint corners[4];
@@ -76,7 +76,7 @@ using namespace WhirlyKit;
     corners[1] = CGPointMake(view.frame.size.width, 0.0);
     corners[2] = CGPointMake(view.frame.size.width, view.frame.size.height);
     corners[3] = CGPointMake(0.0, view.frame.size.height);
-    Point3f planePts[4];
+    Point3d planePts[4];
     bool isValid = true;
     for (unsigned int ii=0;ii<4;ii++)
     {
@@ -107,11 +107,11 @@ using namespace WhirlyKit;
 			break;
 		case UIGestureRecognizerStateChanged:
         {
-            Point3f curLoc = mapView.loc;
-            float newZ = startZ/pinch.scale;
+            Point3d curLoc = mapView.loc;
+            double newZ = startZ/pinch.scale;
             if (minZoom >= maxZoom || (minZoom < newZ && newZ < maxZoom))
             {
-                [mapView setLoc:Point3f(curLoc.x(),curLoc.y(),newZ)];
+                [mapView setLoc:Point3d(curLoc.x(),curLoc.y(),newZ)];
                 if (![self withinBounds:mapView.loc view:glView renderer:sceneRenderer])
                     [mapView setLoc:curLoc];
             }
